fix(ScrollAction): unsafe null assertion operation
This conditional most likely was meant to use || instead of && operator
to check if notificationContext is null OR if no Scrollable was found
for this notificationContext. Using the AND operator instead causes the
null assertion to fail when notificationContext is null, while also
causing an assertion failure when notificationContext is not null, but
no Scrollable is found for it. While at it, let's also deduplicate the
Scrollable.maybeOf call and remove the unnecessary assertion.